extracted common repository logic into QuerydslExpressionFactory.java
