+++ infobip-spring-data-r2dbc-querydsl/src/main/java/com/infobip/spring/data/r2dbc/QuerydslR2dbcRepositoryFactory.java
@@ -15,13 +15,11 @@
  */
 package com.infobip.spring.data.r2dbc;
 
-import com.google.common.base.CaseFormat;
 import com.infobip.spring.data.common.Querydsl;
-import com.querydsl.core.types.*;
+import com.infobip.spring.data.common.QuerydslExpressionFactory;
+import com.querydsl.core.types.ConstructorExpression;
 import com.querydsl.core.types.dsl.PathBuilder;
 import com.querydsl.sql.*;
-import org.springframework.core.ResolvableType;
-import org.springframework.data.annotation.PersistenceConstructor;
 import org.springframework.data.r2dbc.convert.R2dbcConverter;
 import org.springframework.data.r2dbc.core.R2dbcEntityOperations;
 import org.springframework.data.r2dbc.repository.support.R2dbcRepositoryFactory;
@@ -30,14 +28,6 @@ import org.springframework.data.repository.core.support.RepositoryComposition;
 import org.springframework.data.repository.core.support.RepositoryFragment;
 import org.springframework.r2dbc.core.DatabaseClient;
 import org.springframework.transaction.ReactiveTransactionManager;
-import org.springframework.util.ReflectionUtils;
-
-import javax.annotation.Nullable;
-import java.lang.reflect.*;
-import java.util.*;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 public class QuerydslR2dbcRepositoryFactory extends R2dbcRepositoryFactory {
 
@@ -45,6 +35,8 @@ public class QuerydslR2dbcRepositoryFactory extends R2dbcRepositoryFactory {
     private final ReactiveTransactionManager reactiveTransactionManager;
     private final DatabaseClient databaseClient;
     private final R2dbcConverter converter;
+    private final QuerydslExpressionFactory querydslExpressionFactory = new QuerydslExpressionFactory(
+            QuerydslR2dbcFragment.class);
 
     public QuerydslR2dbcRepositoryFactory(R2dbcEntityOperations operations,
                                           SQLQueryFactory sqlQueryFactory,
@@ -61,9 +53,10 @@ public class QuerydslR2dbcRepositoryFactory extends R2dbcRepositoryFactory {
     protected RepositoryComposition.RepositoryFragments getRepositoryFragments(RepositoryMetadata metadata) {
 
         RepositoryComposition.RepositoryFragments fragments = super.getRepositoryFragments(metadata);
-        RelationalPathBase<?> path = getRelationalPathBaseFromQueryRepositoryClass(metadata.getRepositoryInterface());
+        RelationalPathBase<?> path = querydslExpressionFactory.getRelationalPathBaseFromQueryRepositoryClass(
+                metadata.getRepositoryInterface());
         Class<?> type = metadata.getDomainType();
-        ConstructorExpression<?> constructorExpression = getConstructorExpression(type, path);
+        ConstructorExpression<?> constructorExpression = querydslExpressionFactory.getConstructorExpression(type, path);
         RepositoryFragment<Object> simpleQuerydslJdbcFragment = createSimpleQuerydslR2dbcFragment(path,
                                                                                                   constructorExpression);
         RepositoryFragment<Object> querydslJdbcPredicateExecutor = createQuerydslJdbcPredicateExecutor(
@@ -95,78 +88,5 @@ public class QuerydslR2dbcRepositoryFactory extends R2dbcRepositoryFactory {
                 databaseClient,
                 converter);
         return RepositoryFragment.implemented(querydslJdbcPredicateExecutor);
-    }
-
-    private ConstructorExpression<?> getConstructorExpression(Class<?> type, RelationalPath<?> pathBase) {
-        Constructor<?> constructor = getConstructor(type);
-
-        if (constructor == null) {
-            throw new IllegalArgumentException(
-                    "Could not discover preferred constructor for " + type);
-        }
-
-        Map<String, Path<?>> columnNameToColumn = pathBase.getColumns()
-                                                          .stream()
-                                                          .collect(Collectors.toMap(
-                                                                  column -> column.getMetadata().getName(),
-                                                                  Function.identity()));
-
-        Path<?>[] paths = Stream.of(constructor.getParameters())
-                                .map(Parameter::getName)
-                                .map(columnNameToColumn::get)
-                                .toArray(Path[]::new);
-
-        return Projections.constructor(type, paths);
-    }
-
-    @Nullable
-    private Constructor<?> getConstructor(Class<?> type) {
-        Constructor<?>[] declaredConstructors = type.getDeclaredConstructors();
-        Constructor<?> persistenceConstructor = Arrays.stream(declaredConstructors)
-                                                      .filter(constructor -> constructor.isAnnotationPresent(
-                                                              PersistenceConstructor.class))
-                                                      .findAny()
-                                                      .orElse(null);
-
-        if (Objects.nonNull(persistenceConstructor)) {
-            return persistenceConstructor;
-        }
-
-        return Arrays.stream(declaredConstructors)
-                     .max(Comparator.comparingInt(Constructor::getParameterCount))
-                     .orElse(null);
-    }
-
-    private RelationalPathBase<?> getRelationalPathBaseFromQueryRepositoryClass(Class<?> repositoryInterface) {
-
-        Class<?> entityType = ResolvableType.forClass(repositoryInterface)
-                                            .as(QuerydslR2dbcFragment.class)
-                                            .getGeneric(0)
-                                            .resolve();
-        if (entityType == null) {
-            throw new IllegalArgumentException("Could not resolve query class for " + repositoryInterface);
-        }
-
-        return getRelationalPathBaseFromQueryClass(getQueryClass(entityType));
-    }
-
-    private Class<?> getQueryClass(Class<?> entityType) {
-        String fullName = entityType.getPackage().getName() + ".Q" + entityType.getSimpleName();
-        try {
-            return entityType.getClassLoader().loadClass(fullName);
-        } catch (ClassNotFoundException e) {
-            throw new IllegalArgumentException("Unable to load class " + fullName);
-        }
-    }
-
-    private RelationalPathBase<?> getRelationalPathBaseFromQueryClass(Class<?> queryClass) {
-        String fieldName = CaseFormat.UPPER_CAMEL.to(CaseFormat.LOWER_CAMEL, queryClass.getSimpleName().substring(1));
-        Field field = ReflectionUtils.findField(queryClass, fieldName);
-
-        if (field == null) {
-            throw new IllegalArgumentException("Did not find a static field of the same type in " + queryClass);
-        }
-
-        return (RelationalPathBase<?>) ReflectionUtils.getField(field, null);
     }
 }
